Added the same modifications to ModifyPart GUI
Added the same modifications to ModifyPart GUI.
initialized the Products Datasource grid equal to the products Binding List. 
Set the selection method to full row.
Made the grid read-only.
Turned off multi-select.
Removed bottom empty row.
Added code to the modifyProduct click. created a var selectedProduct and set it equal to a dynamic cast of Product and then made the current row the DataBoundItem.
In the ModifyPart.cs file and the ModifyProduct.cs, we created another constructor that takes in Part part and Product product. Then set the text from the corresponding textboxes equal to the properties with a ToString method attached.